sqlite: Added transaction mutex to support go routines better.,
This mutex is used for all transactions. I tried to get it working with a for WRITE only mutex but could not. Altough I could not reproduce the problems I had with it in a test, it was not working in the "fylr" application.

So, the solution now is a "lazy" solution, using the mutex for the period of the entire transaction.