TX begin: use a per db wrapper handle mutex instead of global
The per db handler mutex avoids race problems SQLPRO had when used with multiple Postgres connection pools.

The mutex to protect the TX begin (in write mode) is really only needed for SQLITE, so changed that and made the mutex it local to the db handle.

This patch removes the possible race problems for Postgres backends and allows multiple sqlite backend to be opened in the same Go binary.