Removed ExecContextExp, added ExecContextRowsAffected
The new ExecContextRowsAffected returns rows affected and the last insert id, whereas the removed ExecContextExp would only return the last insert id matching the affected row count with an expected number.

The func gives the programmer more flexibility. For expectiation checks this is a little more code on the caller side, but now it is possible to learn how many rows where actually affected. This is helpful if you don't know before how many rows are going to be affected by e.g. an UPDATE statement.

see #69291