Fixed replacements error "replaceArgs: Expecting #2 arg. Got: 1 args"
If the SQL for a query contained a ? inside and escaped string, the argument slice replacements (for IN ?) queries would fail to properly detect such cases. The replacement logic demanded to use ?? and @@ (for name replacements) in such cases.

The solution now properly parses sql strings (escaped using ' or ") and removes the escaped ?? or @@ support.